* gcc.dg/vect/vect-117.c (main1): Avoid pointer to integer         cast warning.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/autovect-branch@107625 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vect/vect-117.c
@@ -10,8 +10,8 @@
 int main1 ()
 {
   int i;
-  int** p;
-  int* ia[N];
+  size_t ** p;
+  size_t * ia[N];
 
   /* Unknown evolution.  */
   for (i = 0; i < N; i++)
